# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API",
        "description": "Документация сервиса.",
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "api/album/": {
            "get": {
                "tags": [
                    "Album"
                ],
                "summary": "Альбомы",
                "description": "Возвращаются все альбомы",
                "operationId": "78e73d394629fa0e41c3d0f2939743f6",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Album"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "Album"
                ],
                "summary": "Альбомы",
                "description": "Создаёт новый альбом",
                "operationId": "ecfee6a36c304a1c4fb56a3e276dacdc",
                "parameters": [
                    {
                        "name": "track_id",
                        "in": "query",
                        "description": "Id песен",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Название альбома",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "release",
                        "in": "query",
                        "description": "Дата издания",
                        "required": false,
                        "schema": {
                            "type": "date"
                        }
                    },
                    {
                        "name": "Author_id",
                        "in": "query",
                        "description": "Id Автора",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "order_number",
                        "in": "query",
                        "description": "Порядковые номера песен",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "api/album/{id}": {
            "put": {
                "tags": [
                    "Album"
                ],
                "summary": "Альбомы",
                "description": "Создаёт новый альбом",
                "operationId": "7fa928a951eb3ce88e43cc00bc6a6be6",
                "parameters": [
                    {
                        "name": "track_id",
                        "in": "query",
                        "description": "Id песен",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Название альбома",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "release",
                        "in": "query",
                        "description": "Дата издания",
                        "required": false,
                        "schema": {
                            "type": "date"
                        }
                    },
                    {
                        "name": "Author_id",
                        "in": "query",
                        "description": "Id Автора",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "order_number",
                        "in": "query",
                        "description": "Порядковые номера песен",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Album"
                ],
                "summary": "Альбомы",
                "description": "Создаёт новый альбом",
                "operationId": "ebfdeadd046d2a28784dd0bdacb91a02",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Album": {
                "title": "Album",
                "description": "Album model",
                "properties": {
                    "title": {
                        "title": "title",
                        "description": "Название альбома",
                        "type": "string",
                        "format": "string",
                        "example": "Lorelie"
                    },
                    "release": {
                        "title": "release",
                        "description": "Дата выхода альбома",
                        "type": "string",
                        "format": "date",
                        "example": "1992-01-20"
                    },
                    "author_id": {
                        "title": "author_id",
                        "description": "Автор альбома",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Album"
                }
            },
...
        }
    },
...
}